pyproject.toml
[tool.poetry]
name = "chatbot-task-bot"
version = "0.1.0"
description = ""
authors = ["akemalyildirim1 <akemalyildirim1@gmail.com>"]
readme = "README.md"
packages = [
    { include = "src/" },
    {include = "*", from = "src"},
]

[tool.poetry.dependencies]
python = "^3.12"
botbuilder-integration-aiohttp = "^4.15.1"
pydantic = "^2.7.4"
pydantic-settings = "^2.3.3"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.10"
mypy = "^1.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
line-length = 88
exclude = ["tests"]

[tool.ruff.lint]
select = [
    # pydocstyle
    "D",
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
plugins = ["pydantic.mypy"]
ignore_missing_imports = true
strict_optional = false